Move AHP components inside admin folder
