+++ frontend/src/components/admin/ahp/AhpComparison.tsx
